Partially implement resolver
After implementing the resolver and wrestling with squaring the differences
in how scopes are handled between it and the interpreter, I hit a block in
differences in how Java and Rust hash keys for Maps. From testing, it appears
the Java implementation uses some information about the instance of a class
in the hashing, which by default Rust does not, using only the values of
the struct. Additionally, due to how I've chosen to implement expressions,
we've lost the ability to diambiguate between two discrete instances of the
same expression on the same line.

However, the 'bad case' example used in the book executes correctly without
the resolver in place. I'm still unsure how my implementation of a stack based
Environment differs in practice from the recursive Environment in the book.
I suspect it is due to how I chose to implement the new/end scope vs the
behavior of the Java implementation having (I think) multiple mutable references
to the environment, but my understanding of that language is limited so
this remains only a theory.

For now, I'm going to park this partially implemented and move on to the next
section, to revisit this if needed after implmenting Classes.

Signed-off-by: Scott Macfarlane <macfarlane.scott@gmail.com>